processes: Rewrite process->script.
This is a first pass at simplifying process->script by using fewer
monadic values and procedures, and by reducing the overall complexity.

* gwl/processes.scm (containerize): Do not wrap in
with-imported-modules, because we make all of Guix available in the
environment.
(process->script): Remove with-imported-modules and add all Guix
modules to the load path instead; use lowerable profile record instead
of profile->derivation; use program-file and computed-file instead of
gexp->script and gexp->derivation.