Remove guix from default environment.
This is a follow-up to commit
f2ec30d2e806db81155a0f8d7a51be699d4bafbd.

Including *all* of Guix was not a good idea after all.  Yes, it does
remove the need for compilation of modules, but it dramatically
increases the size of the environment by several hundred megabytes,
which is not always acceptable.  Besides, compiling the needed subset
of modules only needs to be done once (until the next GC).