Compute workflow process order and scripts at once.
Much of the complexity of running and preparing workflows stems from
the fact that we compute values on demand.  Sometimes this requires
passing down values from one procedure to another, again and again,
just to use the value in a single procedure at the end of the call
chain.  I prefer my spaghetti on a plate, not in my code.

With this commit we compute the process order, the process scripts,
the process wrapper scripts, and the cache prefix all at once and
store them in a <computed-workflow> record.  There is still some
redundancy (the record contains a hash table and a procedure to let
callers find the name of the script for a given process), but it
allows us to pass much fewer values from procedure to procedure.

* gwl/workflows.scm (<computed-workflow>): New record.
(compute-workflow, computed-workflow-workflow,
computed-workflow-script-proc, computed-workflow-ordered-processes):
New procedures.
(workflow-prepare): Use compute-workflow; accept inputs.
(workflow-run): Use compute-workflow; simplify internal RUN procedure...
(workflow-kons): ...which allows us to simplify this procedure.
* gwl/workflows/utils.scm (script-name): Remove procedure.
* gwl/processes.scm (process->script): Merge inner and outer
procedure; only compute wrappers when new argument SCRIPTS-TABLE is
provided; pass that argument to the wrapper generators...
* gwl/process-engines/grid-engine.scm (grid-engine-wrapper): ...like
this one.  Pass the SCRIPTS-TABLE to...
(process->grid-engine-restrictions-string): ...this procedure, which
passes it to...
(process-job-name): ...this procedure to generate unique Grid Engine
job names.
* gwl/cache.scm (process->hash-string): New procedure.
(make-process->cache-prefix): Take SCRIPTS-TABLE instead of
MAKE-SCRIPT; pass it to...
(workflow->data-hashes): ...this procedure, which passes it to...
(process->hash): ...this newly exported procedure; take a
SCRIPTS-TABLE instead of MAKE-SCRIPT and WORKFLOW.
* tests/cache.scm: Adjust to using the SCRIPTS-TABLE instead of
MAKE-SCRIPT.