workflows: Compute fewer scripts.
Computing scripts is expensive, because we need to generate a manifest
over an inferior Guix.  Different instantiations of a process template
were computed individually, even though their scripts are identical.
In addition to mapping processes to their scripts, create a table of
process procedures to process scripts, so that processes that have the
same procedure don't need to have their scripts computed again.

* gwl/workflows.scm (compute-workflow)[scripts-by-process]: Reuse
computed scripts for processes with identical procedures.