Launch a new Guile process when starting GWL.
See this mailing list thread for context:
https://lists.gnu.org/archive/html/gwl-devel/2022-02/msg00000.html

Although we try to reset %load-path and %load-compiled-path
immediately when execution is passed to the Guix extension the Guile
process will have already loaded some Guix modules.  This leads to a
weird mismatch between the "library Guix" we use in GWL and the
profile hooks when building Guix profiles.

To avoid this confusion we launch a completely new Guile process with
our modified load paths, a process that is completely detached from
the invoking Guix.  This is not pretty and really stretches the
meaning of the term "Guix extension", but it is necessary if we want
to keep a fixed variant of Guix as a library.

* guix/extensions/workflow.scm.in (guix-workflow): Launch Guile after
setting new GUILE_LOAD_PATH and GUILE_LOAD_COMPILED_PATH.  Set
_GWL_INVOKING_GUIX here instead of later.
* gwl/main.scm (guix-workflow-main): Do not set _GWL_INVOKING_GUIX
here; use ARGS instead of looking up COMMAND-LINE directly.
* Makefile.am: Substitute @GUILE@ placeholder.